refactor(creature_dgm_h): consolidate 11 files → 4
Merge 6 experiment files into unified experiments.py:
  - experiment_weight_topology.py   → experiments.py weight
  - experiment_activation_topology.py → experiments.py activation
  - experiment_sequence_topology.py → experiments.py sequence
  - experiment_basin_geometry.py    → experiments.py basin
  - sgd_ablation.py                 → experiments.py sgd
  - experiment_analysis.py          → experiments.py analyze

Absorb __main__.py into __init__.py (python -m still works).
Rewrite README.md to reflect new structure.

Shared helpers deduplicated: load_corpus, cluster_texts,
farthest_point_sample, pca_project, compute_topology,
adam_step, sgd_step, weight vector utilities.

vybn.py untouched. Data dirs (archive/, experiment_results/) untouched.
