+++ Vybn_Mind/creature_dgm_h/README.md
@@ -1,12 +1,23 @@
 # creature_dgm_h
 
-Rotor-modulated character-level prediction. Single file: `vybn.py`.
+Rotor-modulated character-level prediction.
 
 Cl(3,0) geometric algebra computes a rotor from embedding trajectories
 (Pancharatnam phase). The rotor modulates gradient updates: parameters
 aligned with the bivector plane get amplified, orthogonal ones get
 dampened. Standard backprop is the special case where the rotor is
 identity.
+
+## Files
+
+```
+vybn.py              # engine — the creature itself
+experiments.py       # unified probe suite (6 experiments)
+__init__.py          # package init + `python -m` hook
+README.md            # this file
+archive/             # persistent data
+experiment_results/  # output from experiment runs
+```
 
 ## Usage
 
@@ -18,40 +29,27 @@ python -m Vybn_Mind.creature_dgm_h status
 python -m Vybn_Mind.creature_dgm_h audit
 ```
 
-## Topology experiments
+## Experiments
 
-Two approaches for measuring whether text selection affects the topological
-structure of learning (superseding the original raw weight-space experiment,
-which produced a uniform null result due to the curse of dimensionality):
-
-### PCA-first persistence
-
-Projects weight-vector snapshots to ~20 dimensions via PCA before computing
-persistent homology. Concentrates variance along the learning trajectory.
+Six probes characterising the creature's geometry, all in `experiments.py`:
 
 ```bash
-python experiment_weight_topology.py              # full experiment
-python experiment_weight_topology.py --quick      # smoke test
-python experiment_weight_topology.py --pca_dim 30 # custom PCA dim
+python experiments.py weight      [--quick] [--pca_dim 30]
+python experiments.py activation  [--quick]
+python experiments.py sequence    [--quick] [--generations 10]
+python experiments.py basin       [--quick] [--agents 8]
+python experiments.py sgd         [--seeds 5]
+python experiments.py analyze     [--experiment pca|activation|both]
 ```
 
-### Activation-space persistence
-
-Captures hidden-layer activations (16-dim) during training using a fixed probe
-sentence. Lower-dimensional and semantically tied to model behaviour.
-
-```bash
-python experiment_activation_topology.py          # full experiment
-python experiment_activation_topology.py --quick  # smoke test
-```
-
-### Analysis (unified)
-
-```bash
-python experiment_analysis.py                         # both experiments
-python experiment_analysis.py --experiment pca        # PCA only
-python experiment_analysis.py --experiment activation # activation only
-```
+| Probe        | What it measures |
+|:-------------|:-----------------|
+| `weight`     | PCA-first persistent homology on weight-vector snapshots |
+| `activation` | Persistent homology on hidden-layer activation snapshots |
+| `sequence`   | Natural motion recorder (null fitness, uncapped generation) |
+| `basin`      | Loss-landscape geometry around the weight-norm fixed point |
+| `sgd`        | SGD vs Adam ablation on the weight-norm attractor |
+| `analyze`    | Post-hoc statistical analysis of topology results |
 
 ## Dependencies
