--- Vybn_Mind/creature_dgm_h/__main__.py
@@ -1,2 +1,0 @@
-from .vybn import main
-main()
